Switch from QNetwork to libcurl
Originally I implemented this because I was looking for a bug. Turns out
the binary is 7 KB smaller when using libcurl. Also, with libcurl, we can
repeatedly call curl_easy_perform() on the same handle, so we don’t need
to initialize the parameters all over again. Should save some CPU cycles,
memory and power.